+++ src/primitive_traits.rs
@@ -60,7 +60,6 @@ pub trait PrimitiveFloat:
     + FloatVectorBackend<3, Unaligned>
     + FloatVectorBackend<4, Unaligned>
     + FloatUtils
-    + num_primitive::PrimitiveFloat<Bits = <Self as PrimitiveFloat>::Bits>
 {
     /// The unsigned integer type with equal width.
     type Bits: PrimitiveUnsigned;
@@ -125,7 +124,6 @@ pub trait PrimitiveInteger:
     + IntegerVectorBackend<2, Unaligned>
     + IntegerVectorBackend<3, Unaligned>
     + IntegerVectorBackend<4, Unaligned>
-    + num_primitive::PrimitiveInteger
 {
 }
 
@@ -168,7 +166,6 @@ pub trait PrimitiveSigned:
     + SignedVectorBackend<2, Unaligned>
     + SignedVectorBackend<3, Unaligned>
     + SignedVectorBackend<4, Unaligned>
-    + num_primitive::PrimitiveSigned<Unsigned = <Self as PrimitiveSigned>::Unsigned>
 {
     /// The unsigned integer type with equal width.
     type Unsigned: PrimitiveUnsigned;
@@ -186,11 +183,7 @@ pub trait PrimitiveSigned:
 /// their names conflict with floating-point functions. When the type system
 /// allows this, all functions will be available.
 #[expect(private_bounds)]
-pub trait PrimitiveUnsigned:
-    Sealed
-    + PrimitiveInteger
-    + num_primitive::PrimitiveUnsigned<Signed = <Self as PrimitiveUnsigned>::Signed>
-{
+pub trait PrimitiveUnsigned: Sealed + PrimitiveInteger {
     /// The signed integer type with equal width.
     type Signed: PrimitiveSigned;
 }
@@ -269,71 +262,3 @@ impl Sealed for u32 {}
 impl Sealed for u64 {}
 impl Sealed for u128 {}
 impl Sealed for usize {}
-
-/// A module that contains dummy traits for when the `num-primitive` feature
-/// flag is disabled.
-///
-/// This will not be necessary once cfg on trait bounds is supported.
-#[cfg(not(feature = "num-primitive"))]
-mod num_primitive {
-    pub trait PrimitiveFloat {
-        type Bits;
-    }
-
-    pub trait PrimitiveInteger {}
-
-    pub trait PrimitiveSigned {
-        type Unsigned;
-    }
-
-    pub trait PrimitiveUnsigned {
-        type Signed;
-    }
-
-    impl PrimitiveFloat for f32 {
-        type Bits = u32;
-    }
-    impl PrimitiveFloat for f64 {
-        type Bits = u64;
-    }
-
-    impl<T> PrimitiveInteger for T {}
-
-    impl PrimitiveSigned for i8 {
-        type Unsigned = u8;
-    }
-    impl PrimitiveSigned for i16 {
-        type Unsigned = u16;
-    }
-    impl PrimitiveSigned for i32 {
-        type Unsigned = u32;
-    }
-    impl PrimitiveSigned for i64 {
-        type Unsigned = u64;
-    }
-    impl PrimitiveSigned for i128 {
-        type Unsigned = u128;
-    }
-    impl PrimitiveSigned for isize {
-        type Unsigned = usize;
-    }
-
-    impl PrimitiveUnsigned for u8 {
-        type Signed = i8;
-    }
-    impl PrimitiveUnsigned for u16 {
-        type Signed = i16;
-    }
-    impl PrimitiveUnsigned for u32 {
-        type Signed = i32;
-    }
-    impl PrimitiveUnsigned for u64 {
-        type Signed = i64;
-    }
-    impl PrimitiveUnsigned for u128 {
-        type Signed = i128;
-    }
-    impl PrimitiveUnsigned for usize {
-        type Signed = isize;
-    }
-}